Call GetEstablishment on GIAS SOAP api

+++ src/Dfe.Spi.GiasAdapter.Domain/GiasApi/IGiasApiClient.cs
@@ -5,6 +5,6 @@ namespace Dfe.Spi.GiasAdapter.Domain.GiasApi
 {
     public interface IGiasApiClient
     {
-        Task<Establishment> GetEstablishmentAsync(int urn, CancellationToken cancellationToken);
+        Task<Establishment> GetEstablishmentAsync(long urn, CancellationToken cancellationToken);
     }
 }
